requirements.txt
django>=3.2
django_redis
celery==4.4.7
redis==3.5.3
channels>=3.0
channels_redis>=3.0
djangorestframework==3.12.4
psycopg2-binary

